+++ data/aloe.go
@@ -91,7 +91,6 @@ func (*Aloe) writeStage2(c chan string, traits string) {
 
 func (p *Aloe) writeStage3(c chan string, traits string) {
 	c <- fmt.Sprintf(`<block name="plantedAloe3_%s" stage="3" traits="%s" tags="T%dPlant">
-    <drop event="Destroy" name="plantedAloe1_%s" count="1" prob="0.5"/>
     <drop event="Fall" name="resourceYuccaFibers" count="0" prob="1" stick_chance="0"/>
     <drop event="Harvest" name="resourceCropAloeLeaf" count="%d" tag="cropHarvest"/>
     <drop event="Harvest" name="resourceCropAloeLeaf" prob="0.5" count="%d" tag="bonusCropHarvest"/>
@@ -124,9 +123,8 @@ func (p *Aloe) writeStage3(c chan string, traits string) {
 		traits,
 		traits,
 		calculatePlantTier(traits),
-		traits,
 		calculateCropYield(p.CropYield, traits),
 		calculateBonusYield(p.BonusYield, traits),
 		getItemTypeIcon(traits),
-		optionallyAddRenewable(traits, p))
+		getRenewableAndDropTags(traits, p))
 }