+++ data/mushroom.go
@@ -106,7 +106,6 @@ func (*Mushroom) writeStage2(c chan string, traits string) {
 
 func (p *Mushroom) writeStage3(c chan string, traits string) {
 	c <- fmt.Sprintf(`<block name="plantedMushroom3_%s" stage="3" traits="%s" tags="T%dPlant">
-    <drop event="Destroy" name="plantedMushroom1_%s" count="1" prob="0.5"/>
     <drop event="Fall" name="resourceYuccaFibers" count="0" prob="1" stick_chance="0"/>
     <drop event="Harvest" name="foodCropMushrooms" count="%d" tag="cropHarvest"/>
     <drop event="Harvest" name="foodCropMushrooms" prob="0.5" count="%d" tag="bonusCropHarvest"/>
@@ -135,9 +134,8 @@ func (p *Mushroom) writeStage3(c chan string, traits string) {
 		traits,
 		traits,
 		calculatePlantTier(traits),
-		traits,
 		calculateCropYield(p.CropYield, traits),
 		calculateBonusYield(p.BonusYield, traits),
 		getItemTypeIcon(traits),
-		optionallyAddRenewable(traits, p))
+		getRenewableAndDropTags(traits, p))
 }